app/requirements.txt
requests
flask
Flask-SQLAlchemy
mysqlclient
mysql-connector
gunicorn
pydantic

# testing
pytest
pytest-mock